+++ src/Symfony/Component/Security/Core/Authentication/Token/PreAuthenticatedToken.php
@@ -21,24 +21,24 @@ use Symfony\Component\Security\Core\User\UserInterface;
 class PreAuthenticatedToken extends AbstractToken
 {
     private $credentials;
-    private $providerKey;
+    private $firewallName;
 
     /**
      * @param string|\Stringable|UserInterface $user
      * @param mixed                            $credentials
      * @param string[]                         $roles
      */
-    public function __construct($user, $credentials, string $providerKey, array $roles = [])
+    public function __construct($user, $credentials, string $firewallName, array $roles = [])
     {
         parent::__construct($roles);
 
-        if (empty($providerKey)) {
-            throw new \InvalidArgumentException('$providerKey must not be empty.');
+        if ('' === $firewallName) {
+            throw new \InvalidArgumentException('$firewallName must not be empty.');
         }
 
         $this->setUser($user);
         $this->credentials = $credentials;
-        $this->providerKey = $providerKey;
+        $this->firewallName = $firewallName;
 
         if ($roles) {
             $this->setAuthenticated(true);
@@ -49,10 +49,21 @@ class PreAuthenticatedToken extends AbstractToken
      * Returns the provider key.
      *
      * @return string The provider key
+     *
+     * @deprecated since 5.2, use getFirewallName() instead
      */
     public function getProviderKey()
     {
-        return $this->providerKey;
+        if (1 !== \func_num_args() || true !== func_get_arg(0)) {
+            trigger_deprecation('symfony/security-core', '5.2', 'Method "%s" is deprecated, use "getFirewallName()" instead.', __METHOD__);
+        }
+
+        return $this->firewallName;
+    }
+
+    public function getFirewallName(): string
+    {
+        return $this->getProviderKey(true);
     }
 
     /**
@@ -78,7 +89,7 @@ class PreAuthenticatedToken extends AbstractToken
      */
     public function __serialize(): array
     {
-        return [$this->credentials, $this->providerKey, parent::__serialize()];
+        return [$this->credentials, $this->firewallName, parent::__serialize()];
     }
 
     /**
@@ -86,7 +97,7 @@ class PreAuthenticatedToken extends AbstractToken
      */
     public function __unserialize(array $data): void
     {
-        [$this->credentials, $this->providerKey, $parentData] = $data;
+        [$this->credentials, $this->firewallName, $parentData] = $data;
         $parentData = \is_array($parentData) ? $parentData : unserialize($parentData);
         parent::__unserialize($parentData);
     }