+++ src/Symfony/Component/Security/Core/Authentication/Token/SwitchUserToken.php
@@ -23,14 +23,12 @@ class SwitchUserToken extends UsernamePasswordToken
     /**
      * @param string|object $user        The username (like a nickname, email address, etc.), or a UserInterface instance or an object implementing a __toString method
      * @param mixed         $credentials This usually is the password of the user
-     * @param string        $providerKey The provider key
-     * @param string[]      $roles       An array of roles
      *
      * @throws \InvalidArgumentException
      */
-    public function __construct($user, $credentials, string $providerKey, array $roles, TokenInterface $originalToken)
+    public function __construct($user, $credentials, string $firewallName, array $roles, TokenInterface $originalToken)
     {
-        parent::__construct($user, $credentials, $providerKey, $roles);
+        parent::__construct($user, $credentials, $firewallName, $roles);
 
         $this->originalToken = $originalToken;
     }